fix(core): wrong sign for `fundingAsset`
In the spreadsheet, the column designating the amount of fundings in asset units had a different sign than the sign of
the "funding daily" column.
This was replicated in the protocol, but the wrong operator was used to add or subtract this funding to the protocol
balances.

For consistency purposes, the sign of `fundingAsset` was changed to match the sign of `funding`, where a positive sign
means that the long side pays funding to the vault side.