fix(pending)!: store `Validate...` protocol actions in pending actions
It makes much more sense to store the `Validate` kind of an action into
the pending actions queue, because that's the action that is actually
pending. Also, this allows the front-end to directly forward this value
to the oracle middleware to get the correct `validationCost` for this pending
action.

BREAKING CHANGE: the queue `PendingActions` now store `Validate...` protocol actions