refactor(core)!: calculate funding according to trading expos of last state update
BREAKING CHANGE: The `funding` function now only takes the timestamp as input and returns funding and old vault trading expo (as calculated from storage values). The `fundingAsset` function likewise only takes the timestamp as input. The `getLiquidationMultiplier` function also only takes the timestamp as input and reverts if the timestamp is older than the last update timestamp in storage.