test: fix integration tests with forking
The integration tests fixture, when forking is enabled, now rolls the fork to a block that is 1000 blocks in the past, so as to allow `skip`s in the tests while not being in the future. The tests are allowed to skip up to 200 minutes in total to remain in the past.