fix(middleware): don't assume pyth decimals
The Pyth oracle middleware was always assuming 8 decimals for the returned prices, without looking at the `expo` field
of the returned struct. This commit fixes that and also fixes a wrong sign in the mock middleware.

To keep the current `getPythDecimals` function (used in tests notably), we now store the last seen decimals in storage.
We consider a positive `expo` (negative decimals) to not be realistic and so we revert with a custom error in that
case. It was confirmed by the Pyth team that this would only happen if an asset price was in the range of trillions of
trillions, which is not expected.