feat(middleware)!: use cached pyth price (#152)
* fix(actions): safe cast

* refactor(middleware): refactor and fix some bugs in the middlewares

Add interfaces for the inherited contracts, unify integer types to reduce casting

* style(pyth): rename variable

* refactor(middleware): rename some variables

* fix(middleware): don't assume pyth decimals

The Pyth oracle middleware was always assuming 8 decimals for the returned prices, without looking at the `expo` field
of the returned struct. This commit fixes that and also fixes a wrong sign in the mock middleware.

To keep the current `getPythDecimals` function (used in tests notably), we now store the last seen decimals in storage.
We consider a positive `expo` (negative decimals) to not be realistic and so we revert with a custom error in that
case. It was confirmed by the Pyth team that this would only happen if an asset price was in the range of trillions of
trillions, which is not expected.

* feat(middleware)!: return cached pyth price if more recent than chainlink on-chain

Refactored middlewares to avoid code repetition and make intent clearer.

BREAKING CHANGE: removed `getPythDecimals` from oracle middleware

* docs: typos

* refactor(middleware): change error name

* perf(middleware): unchecked math

* fix(middleware): error name

* chore: silence slither false positive

* test(middleware): small refactor and test two new branches in PythOracle

* test(middleware): check cached price retrieval for initiate actions

* test: replace hard-coded values with constants

* test(middleware): update hardcoded pyth data to be more recent and fix integration tests

* test: fix tests

* test: use stETH feed for OracleMiddleware unit tests

* fix(middleware): only accept pyth on-chain price if more recent than `_timeElapsedLimit`

* test: add test for old cached pyth price

* docs(middleware): natspec

* docs(middleware): natspec

* perf(middleware): unchecked math

* test: remove unnecessary casts