refactor: remove sdex and usdn decimals from storage and use a constant instead
Having the USDN and SDEX decimals were not really useful as those will be known in advance and
cannot be changed. So we will use a constant instead named TOKENS_DECIMALS and renove the storage
values and their respective getters as well.

BREAKING CHANGE: `getUsdnDecimals` has been removed, use `TOKENS_DECIMALS` instead, or the
`decimals()` function on the token contract instead