fix(withdraw)!: use USDN shares for withdrawal input and burn (#173)
* fix(withdraw)!: use USDN shares for withdrawal input and burn

The previous implementation assumed that the amount of USDN that was transferred during the initiateWithdrawal remained constant until validateWithdrawal was called, and then burned said amount (stored in the pending action). However, it's possible that a rebase of USDN happens between the two transactions, which would leave some USDN un-burnt in the protocol contract.

This commit modifies the behavior to instead take as input an amount of USDN shares, which is then stored in the pending action (requiring some refactor to utilize more bytes than the original uint128 allowed). The amount of shares is then used to burn the corresponding tokens, which is guaranteed to be exact even if a rebase happened.

The `VaultPendingAction` has been renamed to `DepositPendingAction` and a new `WithdrawalPendingAction` has been created, with more descriptive field names.

Note that the order of the fields in all PendingAction structs changed, the security deposit value is now right after the user address.

BREAKING CHANGE: `initiateWithdrawal` now takes an input amount of USDN shares instead of USDN tokens, and the type is `uint152`. The `VaultPendingAction` type has been replaced by `DepositPendingAction` and `WithdrawalPendingAction`, the `PendingAction` and `LongPendingAction` types have re-ordered fields.

* refactor: add helper functions to split and merge withdrawal amount

* docs: natspec precision

* test: remove unneeded cast

* refactor: rename internal functions

* test: refactor to reduce calculations

* refactor(vault): remove unnecessary cast

* docs: typo