fix(long)!: store liquidation penalty in tick info and use that value for later calculations (#183)
* feat: work in progress

BREAKING CHANGE: `getCurrentTotalExpoByTick` was doing the same as `getTotalExpoByTick` and has been removed, `getCurrentPositionsInTick` was doing the same as `getPositionsInTick` and has been removed. `getTotalExpoByTick` and `getPositionsInTick` have been replaced by `getTickData` which returns a `TickData` struct.

* refactor(long)!: assign liquidation penalty to tick upon first position insertion, and use this value for later calculations

BREAKING CHANGE: `getLiquidationPenalty`, `setLiquidationPenalty` and the `LiquidationPenaltyUpdated` event now all use `uint8` for the value. The `getLongPosition` function now returns a second parameter with the liquidation penalty value for the tick containing this position. On creation of a new long position, the liquidation penalty for that position will not necessarily be the value obtained by `getLiquidationPenalty`, but might be the value stored in the tick that will hold that position.

* feat(long): new function to get a tick's liquidation penalty

* refactor: change return type for `getTickLiquidationPenalty`

* fix: stack too deep

* fix: stack too deep

* test: fix stack too deep

* test(open): add test for stored liquidation penalty

* test(long): add tests for `getTickLiquidationPenalty`

* refactor(actions): move function

* refactor(actions): remove unused variable assignation and improve comments

* perf(actions): remove duplicate code

* ci: action was fixed upstream

* test(open): add test case for new max leverage logic

* test(long): add tests for `getTickLiquidationPenalty`