refactor(open): updating pos liq price in `validateOpenPosition` when above max leverage
The new `_updateLiquidationPrice` function had a lot of code that was duplicated from `_removeAmountFromPosition` and `_saveNewPosition`.
To fix the bug that was previously found, `_saveNewPosition` now doesn't mutate the `_balanceLong` storage variable anymore.
This way of doing the tick update for the position is slightly less efficient than the custom `_updateLiquidationPrice` function, but since it should almost never get used (it's a rare edge case where the leverage goes above the limit between the initiate and validate actions), the reduced deployed bytecode size is preferred.