feat(usdn): add callback mechanism after rebase
Whenever a rebase happens due to a call to `rebase`, a callback handler
contract is called with the old and new divisor values, allowing to
perform additional actions like calling `sync` on a DEX pair. If the
handler address is set to zero, then no callback is made. The callback
handler can return arbitrary data that is forwarded to the caller of
`rebase`.

The rebase function now returns a boolean to indicate whether a rebase
happened, as well as the old divisor value and any callback handler data
(if a rebase happened and a handler was set). This does not constitute a
breaking change.