fix(usdn): decrease allowance even if token amount is zero (#248)
* fix(usdn): decrease allowance even if token amount is zero

Previously, shares were always converted to tokens by rounding to the
nearest integer. If transferring shares that equal to an amount less
than 0.5 wei of tokens, the allowance would previously not be
decremented. This is now fixed by rounding up when calculating the
amount of tokens to remove from the allowance. The conversion function
has been refactor and optimized to be more efficient and allow control
over the rounding.

* test(usdn): add test for allowance update during transferSharesFrom

* test(usdn): add test for allowance update during burnSharesFrom

* test(usdn): add tests for _converToTokens

* feat(usdn): when rounding to closest, round up if at the half-way point

* perf(usdn): avoid double SLOAD

* docs(natspec): clarification

* refactor(usdn): avoid assembly

* test(usdn): make handler function pure

* chore(nix): update flake