feat(usdn)!: add rebase callback (#253)
BREAKING CHANGE: The `LiquidationRewardsManager.getLiquidationRewards` function has an additional argument `rebaseCallbackResult` of type `bytes`.

* fix(usdn): decrease allowance even if token amount is zero

Previously, shares were always converted to tokens by rounding to the
nearest integer. If transferring shares that equal to an amount less
than 0.5 wei of tokens, the allowance would previously not be
decremented. This is now fixed by rounding up when calculating the
amount of tokens to remove from the allowance. The conversion function
has been refactor and optimized to be more efficient and allow control
over the rounding.

* test(usdn): add test for allowance update during transferSharesFrom

* test(usdn): add test for allowance update during burnSharesFrom

* test(usdn): add tests for _converToTokens

* feat(usdn): when rounding to closest, round up if at the half-way point

* perf(usdn): avoid double SLOAD

* feat(usdn): add callback mechanism after rebase

Whenever a rebase happens due to a call to `rebase`, a callback handler
contract is called with the old and new divisor values, allowing to
perform additional actions like calling `sync` on a DEX pair. If the
handler address is set to zero, then no callback is made. The callback
handler can return arbitrary data that is forwarded to the caller of
`rebase`.

The rebase function now returns a boolean to indicate whether a rebase
happened, as well as the old divisor value and any callback handler data
(if a rebase happened and a handler was set). This does not constitute a
breaking change.

* docs(usdn): natspec clarification

* refactor(vault): use new return value to check whether a rebase happened

* test(usdn): add rebase handler test

* docs(natspec): clarification

* refactor(usdn): avoid assembly

* test(usdn): add test for admin function

* test(usdn): make handler function pure

* chore(nix): update flake

* refactor(usdn): remove virtual keyword

* test: move callback event to utils

* refactor(rewards): forward the rebase callback data to the rewards manager

* docs: typo

* refactor: perform rebase in a try-catch block to avoid DoS

* test: assert false

* chore: fix abi export

* test: fix import