feat(rewards): add protocol action to the rewards manager input
This will allow to differentiate between a standalone liquidation vs
liquidation happening during the course of a normal user action. The
interface was refactored into two. IBaseLiquidationRewardsManager now
only exposes the function that the protocol should call. All other
functions that are not required for the protocol are exposed in the
regular ILiquidationRewardsManager interface.