feat(actions)!: add validator argument to `initiateClosePosition`
The owner of a position can now delegate the validation of the position
close to another address.

BREAKING CHANGE: `initiateClosePosition` has a new `validator` argument,
`validateClosePosition` now expects the validator as first argument, the
`InitiatedClosePosition` event has an additional `validator` parameter,
the `ValidatedClosePosition` event now reports the validator address
instead of the owner.