feat(middleware): add sanity check on Pyth price timestamp
In the case where Pyth is down at the `targetTimestamp`, we allow the
next available price to be considered valid. However, we should still
check that this timestamp is not too far in the future, so we consider
at the moment that 20 min is the limit for how late the price can be.