fix(rebalancer)!: disallow choosing the validator address on closing
Users of the rebalancer should be allowed to fully close their position, even if that leaves the rebalancer position below the `_minLongAmount` of the protocol. To perform this check, we were previously looking at the rebalancer deposit of the `to` address. This left the protocol subject to an attack whereby a user of the rebalancer could indicate the address of another rebalancer user as the `to` to change the way the checks are performed. Now, we force the `validator` to be the rebalancer user's address, and we check the close amount condition by checking the validator deposit inside the rebalancer.