fix(open): reflect the position fees immediately on the balances
When a new long position is opened, the entry price is raised according to the position fee so that the position is immediately in slight loss. Previously, we were crediting the full value of the collateral to the long side, which was wrong. Now, we calculate what is the position value according to the provided price (without fee) and credit the rest (fee) to the vault instead. During the `validateOpenPosition`, we must take care to cancel any potential PnL from the temporary position and restore the balances according to its final entry value, with the newly provided price.