feat!: more granularity for liquidation penalty
It is now possible to set the liquidation penalty to an arbitrary number of ticks, regardless of the tick spacing.

BREAKING CHANGE: the liquidation penalty is now expressed in ticks and is stored as a uint24. This affects the `LiquidationPenaltyUpdated` event, functions `getLiquidationPenalty`, `setLiquidationPenalty`, `getLongPosition`, `getTickLiquidationPenalty`, and the `TickDat` struct