refactor(middleware): better casting route between i64 and uint256
The previous implementation was casting int64 to uint64 before casting to uint256. In principle, it's better to first cast to int256 and then uint256, as that would allow it to work even if pyth someday decides to return a larger type without warning us. This was changed out of an abundance of caution